refactored clientside classes in separate files
