+++ Core/OfficeDevPnP.Core/OfficeDevPnP.Core.csproj
@@ -559,9 +559,18 @@
     <Compile Include="Framework\Provisioning\Providers\Xml\XMLPnPSchemaV201805Serializer.cs" />
     <Compile Include="Framework\Provisioning\Providers\Xml\XMLPnPSchemaV201801Serializer.cs" />
     <Compile Include="Framework\Provisioning\Providers\Xml\XMLPnPSchemaV201705Serializer.cs" />
-    <Compile Include="Pages\ClientSidePageCanvas.cs" />
-    <Compile Include="Pages\ClientSidePageControls.cs" />
-    <Compile Include="Pages\ClientSidePagesControlData.cs" />
+    <Compile Include="Pages\AvailableClientSideComponents.cs" />
+    <Compile Include="Pages\CanvasColumn.cs" />
+    <Compile Include="Pages\CanvasControl.cs" />
+    <Compile Include="Pages\CanvasSection.cs" />
+    <Compile Include="Pages\CanvasSectionTemplate.cs" />
+    <Compile Include="Pages\ClientSideCanvasControlData.cs" />
+    <Compile Include="Pages\ClientSideCanvasControlPosition.cs" />
+    <Compile Include="Pages\ClientSideCanvasData.cs" />
+    <Compile Include="Pages\ClientSideCanvasPosition.cs" />
+    <Compile Include="Pages\ClientSidePage.cs" />
+    <Compile Include="Pages\ClientSideComponent.cs" />
+    <Compile Include="Pages\ClientSidePageLayoutType.cs" />
     <Compile Include="Entities\UnifiedGroupEntity.cs" />
     <Compile Include="Extensions\ExceptionExtensions.cs" />
     <Compile Include="Framework\Graph\PnPHttpProvider.cs" />
@@ -900,6 +909,13 @@
     <Compile Include="Extensions\PnPClientContext.cs" />
     <Compile Include="IdentityModel\TokenProviders\ADFS\CertificateMixed.cs" />
     <Compile Include="IdentityModel\WSTrustBindings\CertificateWSTrustBinding.cs" />
+    <Compile Include="Pages\ClientSideText.cs" />
+    <Compile Include="Pages\ClientSideTextControlData.cs" />
+    <Compile Include="Pages\ClientSideWebPart.cs" />
+    <Compile Include="Pages\ClientSideWebPartControlData.cs" />
+    <Compile Include="Pages\ClientSideWebPartData.cs" />
+    <Compile Include="Pages\DefaultClientSideWebParts.cs" />
+    <Compile Include="Pages\PromotedState.cs" />
     <Compile Include="Sites\SiteCollection.cs" />
     <Compile Include="Sites\SiteCollectionCreationInformation.cs" />
     <Compile Include="Utilities\Async\SynchronizationContextRemover.cs" />
