+++ Core/OfficeDevPnP.Core/Pages/ClientSideTextControlData.cs
@@ -1,0 +1,19 @@
+﻿using Newtonsoft.Json;
+
+namespace OfficeDevPnP.Core.Pages
+{
+#if !ONPREMISES
+
+    /// <summary>
+    /// Control data for controls of type 4 (= text control)
+    /// </summary>
+    public class ClientSideTextControlData : ClientSideCanvasControlData
+    {
+        /// <summary>
+        /// Gets or sets JsonProperty "editorType"
+        /// </summary>
+        [JsonProperty(PropertyName = "editorType")]
+        public string EditorType { get; set; }
+    }
+#endif
+}
