+++ Core/OfficeDevPnP.Core/Pages/DefaultClientSideWebParts.cs
@@ -1,0 +1,115 @@
+﻿namespace OfficeDevPnP.Core.Pages
+{
+#if !ONPREMISES
+    /// <summary>
+    /// List of possible OOB web parts
+    /// </summary>
+    public enum DefaultClientSideWebParts
+    {
+        /// <summary>
+        /// Third party webpart
+        /// </summary>
+        ThirdParty,
+        /// <summary>
+        /// Content Rollup webpart
+        /// </summary>
+        ContentRollup,
+        /// <summary>
+        /// Bing Map webpart
+        /// </summary>
+        BingMap,
+        /// <summary>
+        /// Content Embed webpart
+        /// </summary>
+        ContentEmbed,
+        /// <summary>
+        /// Document Embed webpart
+        /// </summary>
+        DocumentEmbed,
+        /// <summary>
+        /// Image webpart
+        /// </summary>
+        Image,
+        /// <summary>
+        /// Image Gallery webpart
+        /// </summary>
+        ImageGallery,
+        /// <summary>
+        /// Link Preview webpart
+        /// </summary>
+        LinkPreview,
+        /// <summary>
+        /// News Feed webpart
+        /// </summary>
+        NewsFeed,
+        /// <summary>
+        /// News Reel webpart
+        /// </summary>
+        NewsReel,
+        /// <summary>
+        /// PowerBI Report Embed webpart
+        /// </summary>
+        PowerBIReportEmbed,
+        /// <summary>
+        /// Quick Chart webpart
+        /// </summary>
+        QuickChart,
+        /// <summary>
+        /// Site Activity webpart
+        /// </summary>
+        SiteActivity,
+        /// <summary>
+        /// Video Embed webpart 
+        /// </summary>
+        VideoEmbed,
+        /// <summary>
+        /// Yammer Embed webpart
+        /// </summary>
+        YammerEmbed,
+        /// <summary>
+        /// Events webpart
+        /// </summary>
+        Events,
+        /// <summary>
+        /// Group Calendar webpart
+        /// </summary>
+        GroupCalendar,
+        /// <summary>
+        /// Hero webpart
+        /// </summary>
+        Hero,
+        /// <summary>
+        /// List webpart
+        /// </summary>
+        List,
+        /// <summary>
+        /// Page Title webpart
+        /// </summary>
+        PageTitle,
+        /// <summary>
+        /// People webpart
+        /// </summary>
+        People,
+        /// <summary>
+        /// Quick Links webpart
+        /// </summary>
+        QuickLinks,
+        /// <summary>
+        /// Custom Message Region web part
+        /// </summary>
+        CustomMessageRegion,
+        /// <summary>
+        /// Divider web part
+        /// </summary>
+        Divider,
+        /// <summary>
+        /// Microsoft Forms web part
+        /// </summary>
+        MicrosoftForms,
+        /// <summary>
+        /// Spacer web part
+        /// </summary>
+        Spacer
+    }
+#endif
+}
